feat(groups): show applicants for group owners
HDVT-14
