--- src/components/volunteer/VolunteerGroupMembers.tsx
@@ -1,119 +1,0 @@
-import { StackScreenProps } from '@react-navigation/stack';
-import _shuffle from 'lodash/shuffle';
-import React, { useCallback, useEffect, useState } from 'react';
-import { FlatList, TouchableOpacity } from 'react-native';
-import { useQuery } from 'react-query';
-
-import { consts, texts } from '../../config';
-import { isApplicant, isMember, volunteerUserData } from '../../helpers';
-import { QUERY_TYPES } from '../../queries';
-import { groupMembership } from '../../queries/volunteer';
-import { MEMBER_STATUS_TYPES, ScreenName, VolunteerUser } from '../../types';
-import { BoldText } from '../Text';
-import { Wrapper } from '../Wrapper';
-
-import { VolunteerAvatar } from './VolunteerAvatar';
-
-const { ROOT_ROUTE_NAMES } = consts;
-
-const keyExtractor = ({ user: { guid } }: { user: VolunteerUser }) => guid;
-
-const MAX_AVATARS_COUNT = 10;
-
-export const VolunteerGroupMembers = ({
-  groupId,
-  navigation,
-  isGroupMember,
-  setIsGroupMember,
-  setIsGroupApplicant,
-  isRefetching,
-  isSuccessJoin,
-  isSuccessLeave,
-  isSuccessRequest
-}: {
-  groupId: number;
-  navigation: StackScreenProps<any>['navigation'];
-  isGroupMember?: boolean;
-  setIsGroupMember: (isMember: boolean) => void;
-  setIsGroupApplicant: (isApplicant: boolean) => void;
-  isRefetching: boolean;
-  isSuccessJoin: boolean;
-  isSuccessLeave: boolean;
-  isSuccessRequest: boolean;
-}) => {
-  const { data, refetch } = useQuery(['groupMembership', groupId], () => groupMembership(groupId));
-  const [groupMembers, setGroupMembers] = useState<{ user: VolunteerUser }[]>();
-
-  const setMembers = useCallback(async () => {
-    const { currentUserId } = await volunteerUserData();
-
-    const applicants = data?.results?.filter(
-      (applicant: { status: number }) => applicant.status === MEMBER_STATUS_TYPES.APPLICANT
-    );
-
-    setIsGroupApplicant(isApplicant(currentUserId, applicants));
-
-    const members = data?.results?.filter(
-      (member: { status: number }) => member.status === MEMBER_STATUS_TYPES.MEMBER
-    );
-
-    setIsGroupMember(isMember(currentUserId, members));
-
-    const fewShuffledMembers = _shuffle(members?.slice(0, MAX_AVATARS_COUNT + 1)) as [
-      { user: VolunteerUser }
-    ];
-
-    fewShuffledMembers?.length && setGroupMembers(fewShuffledMembers);
-  }, [data]);
-
-  useEffect(() => {
-    setMembers();
-  }, [setMembers]);
-
-  useEffect(() => {
-    refetch();
-  }, [isSuccessJoin, isSuccessLeave, isSuccessRequest]);
-
-  useEffect(() => {
-    // refetch posts when pull to refresh from parent component
-    isRefetching && refetch();
-  }, [isRefetching]);
-
-  if (!groupMembers?.length) return null;
-
-  return (
-    <Wrapper>
-      <TouchableOpacity
-        onPress={() =>
-          navigation.push(ScreenName.VolunteerIndex, {
-            title: texts.volunteer.members,
-            query: QUERY_TYPES.VOLUNTEER.MEMBERS,
-            queryVariables: groupId,
-            rootRouteName: ROOT_ROUTE_NAMES.VOLUNTEER,
-            isGroupMember
-          })
-        }
-        disabled={!isGroupMember}
-      >
-        <BoldText>{texts.volunteer.members}</BoldText>
-        <FlatList
-          keyExtractor={keyExtractor}
-          data={groupMembers}
-          renderItem={({ item, index }: { item: { user: VolunteerUser }; index: number }) => (
-            <VolunteerAvatar
-              {...{
-                item,
-                index,
-                totalCount: data.total,
-                MAX_AVATARS_COUNT
-              }}
-            />
-          )}
-          showsHorizontalScrollIndicator={false}
-          horizontal
-          bounces={false}
-        />
-      </TouchableOpacity>
-    </Wrapper>
-  );
-};
