+++ src/screens/volunteer/VolunteerIndexScreen.tsx
@@ -108,7 +108,9 @@ export const VolunteerIndexScreen = ({ navigation, route }: StackScreenProps<any
           showBackToTop
           openWebScreen={openWebScreen}
         />
-        {((query === QUERY_TYPES.VOLUNTEER.MEMBERS && isGroupMember) ||
+        {(((query === QUERY_TYPES.VOLUNTEER.MEMBERS ||
+          query === QUERY_TYPES.VOLUNTEER.APPLICANTS) &&
+          isGroupMember) ||
           (query === QUERY_TYPES.VOLUNTEER.CALENDAR && isAttendingEvent)) && (
           <Wrapper style={styles.noPaddingBottom}>
             <Button
